josh/clean-up-old-headers * Remove some surplus old headewr components
