+++ src/js/components/index.js
@@ -9,7 +9,7 @@ export {
 export { WeatherIcon } from './weather-icon';
 export { Page } from './page';
 export { GlobalHeader, GlobalNav } from './global-header';
-export { PageHeader, Title, PageTabsOld, TabLink } from './page-header';
+export { TabLink } from './TabLink';
 export { Table, TableDivider } from './Table';
 export {
     StatusIndicator,
